\name{umairpkg-package}
\alias{umairpkg-package}
\alias{umairpkg}
\docType{package}
\title{
\packageTitle{umairpkg}
}
\description{
\packageDescription{umairpkg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{umairpkg}
\packageIndices{umairpkg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{umairpkg}

Maintainer: \packageMaintainer{umairpkg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
